rag-server/requirements.txt
# Core RAG libraries (최소 버전으로 간소화)
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers

# Vector & Embeddings
huggingface_hub
transformers

# Document loaders
pypdf
python-docx
docx2txt
openpyxl
python-multipart

# Web server
fastapi
uvicorn
pydantic

# Utilities
python-dotenv
openai
